新增了密码加密和JWT鉴权

+++ tensquare_user/src/main/java/com/tensquare/user/controller/UserController.java
@@ -1,15 +1,20 @@
 package com.tensquare.user.controller;
+
 import com.tensquare.user.pojo.User;
 import com.tensquare.user.service.UserService;
 import entity.PageResult;
 import entity.Result;
 import entity.StatusCode;
+import io.jsonwebtoken.Claims;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.web.bind.annotation.*;
+import util.JwtUtil;
 
+import javax.servlet.http.HttpServletRequest;
+import java.util.HashMap;
 import java.util.Map;
 /**
  * 控制器层
@@ -26,7 +31,13 @@ public class UserController {
 
 	@Autowired
 	private RedisTemplate redisTemplate;
-	
+
+	@Autowired
+	private JwtUtil jwtUtil;
+
+	@Autowired
+	private HttpServletRequest request;
+
 	/**
 	 * 查询全部数据
 	 * @return
@@ -92,11 +103,42 @@ public class UserController {
 	}
 	
 	/**
-	 * 删除
+	 * 删除,只有管理员有权限删除用户
 	 * @param id
 	 */
 	@RequestMapping(value="/{id}",method= RequestMethod.DELETE)
 	public Result delete(@PathVariable String id ){
+		/*//前后端约定：前端请求微服务时需要添加头信息Authorization ,内容为Bearer+空格+token
+		String authHeader = request.getHeader("Authorization");
+		//如果为空
+		if(authHeader==null){
+			return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+		}
+		//如果不是Bearer 开头的
+		if(!authHeader.startsWith("Bearer ")){
+			return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+		}
+		//截取token
+		String token = authHeader.substring(7);
+		//如果校验的时候令牌过期,会报错,所以要捕获,
+		try {
+			//解析tonken
+			Claims claims = jwtUtil.parseJWT(token);
+			if(claims==null){
+				return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+			}
+			if(!"admin".equals(claims.get("roles"))){
+				return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+			}
+		}catch (Exception e){
+			return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+		}*/
+
+
+		Claims claims = (Claims) request.getAttribute("admin_claims");
+		if(claims==null){
+			return new Result(false,StatusCode.ACCESSERROR,"权限不足");
+		}
 		userService.deleteById(id);
 		return new Result(true,StatusCode.OK,"删除成功");
 	}
@@ -127,6 +169,20 @@ public class UserController {
 		}
 		userService.register(user);
 		return new Result(true,StatusCode.OK,"注册成功");
+	}
+
+	@RequestMapping(value = "/login",method = RequestMethod.POST)
+	public Result login(@RequestBody User loginUser){
+		User user = userService.findByMobileAndPassword(loginUser.getMobile(), loginUser.getPassword());
+		if(user==null){
+			return new Result(false,StatusCode.LOGINERROR,"登录失败");
+		}
+		String token = jwtUtil.createJWT(user.getId(), user.getMobile(), "user");
+		Map<String,Object> map = new HashMap<>();
+		map.put("nickName",user.getNickname());//存储昵称
+		map.put("token",token);
+		map.put("roles","user");
+		return new Result(true,StatusCode.OK,"登陆成功",map);
 	}
 	
 }
